done api

+++ src/main/java/com/tindev/modal/PaymentSummary.java
@@ -3,11 +3,13 @@ package com.tindev.modal;
 import com.tindev.domain.PaymentType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PaymentSummary {
 
     private PaymentType paymentType;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private int transactionCount;
     private double percentage;
 }
